rename reportviews and adjust imports
