+++ toponetx/classes/__init__.py
@@ -4,7 +4,7 @@ from .hyperedge import *
 
 from .node import *
 
-from .reportview import *
+from .reportviews import *
 
 from .complex import *
 from .simplicial_complex import *
