+++ toponetx/classes/cell_complex.py
@@ -22,7 +22,7 @@ from scipy.sparse import csr_matrix
 from toponetx.classes.cell import Cell
 from toponetx.classes.combinatorial_complex import CombinatorialComplex
 from toponetx.classes.complex import Complex
-from toponetx.classes.reportview import CellView
+from toponetx.classes.reportviews import CellView
 from toponetx.exception import TopoNetXError
 
 __all__ = ["CellComplex"]
